QuantPulse-Backend/requirements.txt
# QuantPulse Backend Dependencies
# Core framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# CORS and middleware
python-multipart>=0.0.6

# Environment variables (for future use)
python-dotenv>=1.0.0

# Stock market data
yfinance>=0.2.36

# News API
newsapi-python>=0.2.7
requests>=2.31.0

# Sentiment Analysis
textblob>=0.18.0

# Async HTTP client (for internal API calls)
httpx>=0.27.0
